Rename stronghold module
