+++ identity-account-storage/src/lib.rs
@@ -21,7 +21,7 @@ pub mod error;
 pub mod identity;
 pub mod storage;
 #[cfg(feature = "stronghold")]
-pub mod stronghold;
+pub mod stronghold_old;
 pub mod types;
 pub mod utils;
