+++ identity-account-storage/src/storage/stronghold.rs
@@ -27,13 +27,13 @@ use tokio::sync::RwLockWriteGuard;
 use crate::error::Result;
 use crate::identity::ChainState;
 use crate::storage::Storage;
-use crate::stronghold::default_hint;
-use crate::stronghold::ClientPath;
-use crate::stronghold::Context;
-use crate::stronghold::Snapshot;
-use crate::stronghold::Store;
-use crate::stronghold::StrongholdError;
-use crate::stronghold::Vault;
+use crate::stronghold_old::default_hint;
+use crate::stronghold_old::ClientPath;
+use crate::stronghold_old::Context;
+use crate::stronghold_old::Snapshot;
+use crate::stronghold_old::Store;
+use crate::stronghold_old::StrongholdError;
+use crate::stronghold_old::Vault;
 use crate::types::KeyLocation;
 use crate::types::Signature;
 use crate::utils::derive_encryption_key;
