+++ identity-account-storage/src/stronghold_old/client_path.rs
