+++ identity-account-storage/src/stronghold_old/error.rs
