+++ identity-account-storage/src/stronghold_old/hint.rs
