+++ identity-account-storage/src/stronghold_old/status.rs
